+++ src/main/scala/satify/dsl/Conversion.scala
@@ -1,0 +1,7 @@
+package satify.dsl
+
+import satify.model.Expression.Symbol
+object Conversion:
+
+  given Conversion[String, Symbol] with
+    def apply(s: String): Symbol = Symbol(s)
