--- src/main/scala/satify/dsl/DSLExports.scala
@@ -1,6 +1,0 @@
-package satify.dsl
-
-object DSLExports:
-  export satify.dsl.DSL.{given, *}
-  export satify.dsl.SatEncodings.*
-  export satify.dsl.Numbers.*
